requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
config==0.5.1
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
openpyxl==3.0.9
    # via -r requirements.in
psycopg2==2.9.2
    # via -r requirements.in
python-dotenv==0.19.2
    # via -r requirements.in
